requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
requests==2.32.3
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
tqdm==4.66.5
sentencepiece==0.2.0
protobuf==5.28.2
safetensors==0.4.5
huggingface-hub==0.25.2
transformers==4.44.2
